Move NewLicenseTest to integration tests
The NewLicenseTest class was relocated from the 'controllertests' package to the 'integration' package to better align with its purpose as an integration test. This change improves the organization of the test suite.
